fix: add setRoleAdmin for VAULT role and merkle tree demo script
