+++ scripts/demo/slisBnbWeight.json
@@ -1,0 +1,29 @@
+{
+    "users": [
+      {
+        "account": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
+        "week": 1,
+        "weight": "1"
+      },
+      {
+        "account": "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+        "week": 1,
+        "weight": "2"
+      },
+      {
+        "account": "0x3C44CdDdB6a900fa2b585dd299e03d12FA4293BC",
+        "week": 1,
+        "weight": "3"
+      },
+      {
+        "account": "0x90F79bf6EB2c4f870365E785982E1f101E93b906",
+        "week": 1,
+        "weight": "4"
+      },
+      {
+        "account": "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+        "week": 1,
+        "weight": "5"
+      }
+    ]
+  }
